% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files-validation.R
\name{validate_version}
\alias{validate_version}
\title{Validate Version}
\usage{
validate_version(dictionary_path)
}
\arguments{
\item{dictionary_path}{string containing path to the dictionary directory.}
}
\value{
list with: Dictionary file name, Dictionary version, Boolean representing if the versions match.
}
\description{
Validate dictionary version between file name and summary sheets.
}
